Move homepage app queries to its own folder
