+++ src/routes/Admin-Web-App/appRouter.js
@@ -3,7 +3,6 @@ const router = express.Router()
 const { ProtectAuthRoutes } = require('../../middlewares/auth')
 const { 
     getUserProjects, 
-    getHomePageData,
 } = require('../../controllers/Admin-Web-App/appQueries/projects')
 const {
     getUserFeatures,
@@ -11,6 +10,7 @@ const {
 } = require('../../controllers/Admin-Web-App/appQueries/features')
 const { getVariables } = require('../../controllers/Admin-Web-App/appQueries/variables')
 const { getApiKeys } = require('../../controllers/Admin-Web-App/appQueries/api-keys')
+const { getHomePageData } = require('../../controllers/Admin-Web-App/appQueries/homepage')
 
 
 router.get('/projects', ProtectAuthRoutes, getUserProjects)
